put HpsGblRefitter mille-output into GBLRefitterDriver and fix bin file-writing bugs
